# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"_36_Enums.RequestStatus": {
				"type": "string",
				"enum": [
					"PENDING",
					"ASSIGNED",
					"IN_PROGRESS",
					"COMPLETED",
					"CANCELLED"
				]
			},
...
			"_36_Enums.AdminRole": {
				"type": "string",
				"enum": [
					"SUPER_ADMIN",
					"ADMIN"
				]
			},
...
			"CreateApplicationDocRequest": {
				"properties": {
					"customerName": {
						"type": "string"
					},
					"customerPhone": {
						"type": "string"
					},
					"customerEmail": {
						"type": "string"
					},
					"service": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"documentUrl": {
						"type": "string"
					},
					"preferredDate": {
						"type": "string"
					},
					"tasker": {
						"type": "string"
					}
				},
				"required": [
					"customerName",
					"customerPhone",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateApplicationDocRequest": {
				"properties": {
					"customerName": {
						"type": "string"
					},
					"customerPhone": {
						"type": "string"
					},
					"customerEmail": {
						"type": "string"
					},
					"service": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"documentUrl": {
						"type": "string"
					},
					"preferredDate": {
						"type": "string"
					},
					"tasker": {
						"type": "string"
					}
				},
				"required": [
					"customerName",
					"customerPhone",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateAdminBody": {
				"properties": {
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"ADMIN",
							"SUPER_ADMIN"
						]
					}
				},
				"required": [
					"fullName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateAdminBody": {
				"properties": {
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"ADMIN",
							"SUPER_ADMIN"
						]
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rise-motive-be",
		"version": "1.0.0",
		"description": "Rise Motive Backend API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/application-docs": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"tasker": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/_36_Enums.RequestStatus"
										},
										"preferredDate": {
											"type": "string"
										},
										"documentUrl": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"service": {
											"type": "string"
										},
										"customerEmail": {
											"type": "string"
										},
										"customerPhone": {
											"type": "string"
										},
										"customerName": {
											"type": "string"
										},
										"trackingCode": {
											"type": "string"
										}
									},
									"required": [
										"id",
										"updatedAt",
										"createdAt",
										"tasker",
										"status",
										"preferredDate",
										"documentUrl",
										"description",
										"service",
										"customerEmail",
										"customerPhone",
										"customerName",
										"trackingCode"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Application Documents"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateApplicationDocRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"items": {
											"items": {
												"properties": {
													"id": {
														"type": "number",
														"format": "double"
													},
													"updatedAt": {
														"type": "string",
														"format": "date-time"
													},
													"createdAt": {
														"type": "string",
														"format": "date-time"
													},
													"tasker": {
														"type": "string"
													},
													"status": {
														"$ref": "#/components/schemas/_36_Enums.RequestStatus"
													},
													"preferredDate": {
														"type": "string"
													},
													"documentUrl": {
														"type": "string"
													},
													"description": {
														"type": "string"
													},
													"service": {
														"type": "string"
													},
													"customerEmail": {
														"type": "string"
													},
													"customerPhone": {
														"type": "string"
													},
													"customerName": {
														"type": "string"
													},
													"trackingCode": {
														"type": "string"
													}
												},
												"required": [
													"id",
													"updatedAt",
													"createdAt",
													"tasker",
													"status",
													"preferredDate",
													"documentUrl",
													"description",
													"service",
													"customerEmail",
													"customerPhone",
													"customerName",
													"trackingCode"
												],
												"type": "object"
											},
											"type": "array"
										},
										"total": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"items",
										"total"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Application Documents"
				],
				"security": [],
				"parameters": []
			}
		},
		"/application-docs/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"tasker": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/_36_Enums.RequestStatus"
										},
										"preferredDate": {
											"type": "string"
										},
										"documentUrl": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"service": {
											"type": "string"
										},
										"customerEmail": {
											"type": "string"
										},
										"customerPhone": {
											"type": "string"
										},
										"customerName": {
											"type": "string"
										},
										"trackingCode": {
											"type": "string"
										}
									},
									"required": [
										"id",
										"updatedAt",
										"createdAt",
										"tasker",
										"status",
										"preferredDate",
										"documentUrl",
										"description",
										"service",
										"customerEmail",
										"customerPhone",
										"customerName",
										"trackingCode"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Application Documents"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"tasker": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/_36_Enums.RequestStatus"
										},
										"preferredDate": {
											"type": "string"
										},
										"documentUrl": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"service": {
											"type": "string"
										},
										"customerEmail": {
											"type": "string"
										},
										"customerPhone": {
											"type": "string"
										},
										"customerName": {
											"type": "string"
										},
										"trackingCode": {
											"type": "string"
										}
									},
									"required": [
										"id",
										"updatedAt",
										"createdAt",
										"tasker",
										"status",
										"preferredDate",
										"documentUrl",
										"description",
										"service",
										"customerEmail",
										"customerPhone",
										"customerName",
										"trackingCode"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Application Documents"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateApplicationDocRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"tasker": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/_36_Enums.RequestStatus"
										},
										"preferredDate": {
											"type": "string"
										},
										"documentUrl": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"service": {
											"type": "string"
										},
										"customerEmail": {
											"type": "string"
										},
										"customerPhone": {
											"type": "string"
										},
										"customerName": {
											"type": "string"
										},
										"trackingCode": {
											"type": "string"
										}
									},
									"required": [
										"id",
										"updatedAt",
										"createdAt",
										"tasker",
										"status",
										"preferredDate",
										"documentUrl",
										"description",
										"service",
										"customerEmail",
										"customerPhone",
										"customerName",
										"trackingCode"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Application Documents"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin": {
			"get": {
				"operationId": "GetAllAdmins",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"role": {
												"$ref": "#/components/schemas/_36_Enums.AdminRole"
											},
											"isActive": {
												"type": "boolean"
											},
											"email": {
												"type": "string"
											},
											"fullName": {
												"type": "string"
											},
											"id": {
												"type": "number",
												"format": "double"
											},
											"updatedAt": {
												"type": "string",
												"format": "date-time"
											},
											"createdAt": {
												"type": "string",
												"format": "date-time"
											}
										},
										"required": [
											"role",
											"isActive",
											"email",
											"fullName",
											"id",
											"updatedAt",
											"createdAt"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateAdmin",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"role": {
											"$ref": "#/components/schemas/_36_Enums.AdminRole"
										},
										"isActive": {
											"type": "boolean"
										},
										"email": {
											"type": "string"
										},
										"fullName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"role",
										"isActive",
										"email",
										"fullName",
										"id",
										"updatedAt",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAdminBody"
							}
						}
					}
				}
			}
		},
		"/admin/{id}": {
			"get": {
				"operationId": "GetAdminById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"role": {
											"$ref": "#/components/schemas/_36_Enums.AdminRole"
										},
										"isActive": {
											"type": "boolean"
										},
										"email": {
											"type": "string"
										},
										"fullName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"role",
										"isActive",
										"email",
										"fullName",
										"id",
										"updatedAt",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"role": {
											"$ref": "#/components/schemas/_36_Enums.AdminRole"
										},
										"isActive": {
											"type": "boolean"
										},
										"email": {
											"type": "string"
										},
										"fullName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"role",
										"isActive",
										"email",
										"fullName",
										"id",
										"updatedAt",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAdminBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"lockedUntil": {
											"type": "string",
											"format": "date-time"
										},
										"loginAttempts": {
											"type": "number",
											"format": "double"
										},
										"role": {
											"$ref": "#/components/schemas/_36_Enums.AdminRole"
										},
										"password": {
											"type": "string"
										},
										"profilePicture": {
											"type": "string"
										},
										"isActive": {
											"type": "boolean"
										},
										"email": {
											"type": "string"
										},
										"fullName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"lockedUntil",
										"loginAttempts",
										"role",
										"password",
										"profilePicture",
										"isActive",
										"email",
										"fullName",
										"id",
										"updatedAt",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/activate/{id}": {
			"put": {
				"operationId": "Activate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isActive": {
											"type": "boolean"
										},
										"fullName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"isActive",
										"fullName",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/deactivate/{id}": {
			"put": {
				"operationId": "Deactivate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isActive": {
											"type": "boolean"
										},
										"fullName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"isActive",
										"fullName",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}